@@ -34,7 +34,7 @@
 		"!dist/**/*.test.*"
 	],
 	"peerDependencies": {
-		"svelte": "^3.58.0 || ^4.0.0"
+		"svelte": "^5.0.0"
 	},
 	"repository": {
 		"type": "git",
@@ -42,26 +42,30 @@
 	},
 	"devDependencies": {
 		"@changesets/cli": "^2.26.2",
-		"@sveltejs/adapter-auto": "^2.1.0",
-		"@sveltejs/kit": "^1.27.2",
-		"@sveltejs/package": "^2.2.2",
+		"@sveltejs/adapter-auto": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/package": "^2.3.7",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@testing-library/svelte": "^4.0.4",
+		"@types/eslint": "^9.6.0",
 		"@typescript-eslint/eslint-plugin": "^5.62.0",
 		"@typescript-eslint/parser": "^5.62.0",
-		"eslint": "^8.52.0",
-		"eslint-config-prettier": "^8.10.0",
-		"eslint-plugin-svelte": "^2.34.0",
+		"eslint": "^9.7.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-svelte": "^2.36.0",
+		"globals": "^15.0.0",
 		"happy-dom": "^8.9.0",
 		"jsdom": "^21.1.2",
-		"prettier": "^2.8.8",
-		"prettier-plugin-svelte": "^2.10.1",
+		"prettier": "^3.3.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"publint": "^0.2.5",
-		"svelte": "^4.2.2",
-		"svelte-check": "^3.5.2",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"tslib": "^2.6.2",
-		"typescript": "^5.2.2",
-		"vite": "^4.5.0",
-		"vitest": "^0.34.6"
+		"typescript": "^5.5.0",
+		"typescript-eslint": "^8.0.0",
+		"vite": "^5.4.4",
+		"vitest": "^1.0.0"
 	},
 	"type": "module",
 	"dependencies": {
