@@ -1,7 +1,17 @@
+<script>
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
+</script>
+
 <pre
 	style:white-space="pre"
 	style:word-wrap="break-word"
 	style:overflow-x="auto"
 	style:font-family="monospace">
-<slot />
+{@render children?.()}
 </pre>
