@@ -1,13 +1,17 @@
 <script lang="ts">
 	import { escapeHTML } from 'es-escape-html';
 
-	export let src = '';
-	export let width = '';
-	export let height = '';
-	export let title = '';
+	interface Props {
+		src?: string;
+		width?: string;
+		height?: string;
+		title?: string;
+	}
+
+	let { src = '', width = '', height = '', title = '' }: Props = $props();
 </script>
 
-<!-- svelte-ignore a11y-media-has-caption -->
+<!-- svelte-ignore a11y_media_has_caption -->
 <video src={escapeHTML(src)} controls width={width || '100%'} height={height || '100%'} {title}>
 	<p>
 		Your browser doesn't support HTML5 video. Here is a{' '}
