@@ -1,1 +1,11 @@
-<strong class="text-black"><slot /></strong>
+<script>
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
+</script>
+
+<strong class="text-black">{@render children?.()}</strong>
