@@ -1,1 +1,11 @@
-<section class="bg-white"><slot /></section>
+<script>
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
+</script>
+
+<section class="bg-white">{@render children?.()}</section>
