@@ -1,1 +1,11 @@
-<code style:font-style="italic"><slot /></code>
+<script>
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
+</script>
+
+<code style:font-style="italic">{@render children?.()}</code>
