@@ -1,6 +1,12 @@
 <script>
-	export let src = '';
-	export let altText = '';
+	/**
+	 * @typedef {Object} Props
+	 * @property {string} [src]
+	 * @property {string} [altText]
+	 */
+
+	/** @type {Props} */
+	let { src = '', altText = '' } = $props();
 </script>
 
 <img {src} alt={altText} />
