@@ -1,1 +1,11 @@
-<i class="italic-class" style:color="red"><slot /></i>
+<script>
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
+</script>
+
+<i class="italic-class" style:color="red">{@render children?.()}</i>
