@@ -1,2 +1,12 @@
-<p class="text-white"><slot /></p>
+<script>
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
+</script>
+
+<p class="text-white">{@render children?.()}</p>
 ,
