@@ -1,1 +1,11 @@
-<u role="button"><slot /> test</u>,
+<script>
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
+</script>
+
+<u role="button">{@render children?.()} test</u>,
