@@ -1,11 +1,14 @@
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vitest/config';
 
-export default defineConfig({
+export default defineConfig(({ mode }) => ({
 	plugins: [sveltekit()],
 	test: {
 		include: ['src/**/*.{test,spec}.{js,ts}'],
 		globals: true,
 		environment: 'happy-dom'
+	},
+	resolve: {
+		conditions: mode === 'test' ? ['browser'] : []
 	}
-});
+}));
